basket remove (ajax)

+++ kpk/basketapp/views.py
@@ -1,5 +1,5 @@
 from basketapp.models import CourseBasket
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import render
 from django.urls import reverse
 from mainapp.models import Course
@@ -27,6 +27,9 @@ def add(request, course_id):
 
 
 def remove(request, course_basket_id):
-    item = CourseBasket.objects.get(id=course_basket_id)
-    item.delete()
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+    if request.is_ajax():
+        item = CourseBasket.objects.get(id=course_basket_id)
+        item.delete()
+        # return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        return JsonResponse({'status': 'ok',
+                             'course_basket_id': course_basket_id})
